Integrated Pipe in project and Process class. Updated tests accordingly.

+++ src/process.cpp
@@ -1,4 +1,5 @@
 #include <libpdb/error.hpp>
+#include <libpdb/pipe.hpp>
 #include <libpdb/process.hpp>
 
 #include <sys/ptrace.h>
@@ -7,6 +8,17 @@
 #include <unistd.h>
 
 namespace pdb {
+
+namespace {
+
+void exitWithPerror(Pipe& channel, std::string const& prefix)
+{
+    auto message = prefix + ": " + std::strerror(errno);
+    channel.write(reinterpret_cast<std::byte*>(message.data()), message.size());
+    exit(-1);
+}
+
+} // namespace
 
 StopReason::StopReason(int waitStatus)
 {
@@ -27,12 +39,15 @@ Process::~Process()
 {
     if(m_pid != 0) {
         int status;
-        if(m_state == ProcessState::Running) {
-            kill(m_pid, SIGSTOP);
-            waitpid(m_pid, &status, 0);
+        if(m_isAttached) {
+            if(m_state == ProcessState::Running) {
+                kill(m_pid, SIGSTOP);
+                waitpid(m_pid, &status, 0);
+            }
+            ptrace(PTRACE_DETACH, m_pid, nullptr, nullptr);
+            kill(m_pid, SIGCONT);
         }
-        ptrace(PTRACE_DETACH, m_pid, nullptr, nullptr);
-        kill(m_pid, SIGCONT);
+
         if(m_terminateOnEnd) {
             kill(m_pid, SIGKILL);
             waitpid(m_pid, &status, 0);
@@ -40,24 +55,39 @@ Process::~Process()
     }
 }
 
-std::unique_ptr<Process> Process::launch(std::filesystem::path path)
+std::unique_ptr<Process> Process::launch(std::filesystem::path path, bool debug)
 {
+    Pipe channel(true);
     pid_t pid;
     if((pid = fork()) < 0) {
         Error::sendErrno("fork failed");
     }
 
     if(pid == 0) {
-        if(ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
-            Error::sendErrno("ptrace TRACEME failed");
+        channel.closeRead();
+        if(debug && ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+            exitWithPerror(channel, "ptrace TRACEME failed");
         }
         if(execlp(path.c_str(), path.c_str(), nullptr) < 0) {
-            Error::sendErrno("execlp failed");
+            exitWithPerror(channel, "execlp failed");
         }
     }
 
-    std::unique_ptr<Process> proc(new Process(pid, true));
-    proc->waitOnSignal();
+    channel.closeWrite();
+    auto data = channel.read();
+    channel.closeRead();
+
+    if(data.size() > 0) {
+        waitpid(pid, nullptr, 0);
+        auto chars = reinterpret_cast<char*>(data.data());
+        Error::send(std::string(chars, chars + data.size()));
+    }
+
+    std::unique_ptr<Process> proc(new Process(pid, true, debug));
+
+    if(debug) {
+        proc->waitOnSignal();
+    }
 
     return proc;
 }
@@ -70,7 +100,7 @@ std::unique_ptr<Process> Process::attach(pid_t pid)
     if(ptrace(PTRACE_ATTACH, pid, nullptr, nullptr) < 0) {
         Error::sendErrno("ptrace ATTACH failed");
     }
-    std::unique_ptr<Process> proc(new Process(pid, false));
+    std::unique_ptr<Process> proc(new Process(pid, false, true));
     proc->waitOnSignal();
     return proc;
 }
@@ -83,7 +113,8 @@ void Process::resume()
     m_state = ProcessState::Running;
 }
 
-StopReason Process::waitOnSignal() {
+StopReason Process::waitOnSignal()
+{
     int waitStatus = 0;
     int options    = 0;
     if(waitpid(m_pid, &waitStatus, options) < 0) {
@@ -94,9 +125,10 @@ StopReason Process::waitOnSignal() {
     return reason;
 }
 
-Process::Process(pid_t pid, bool terminateOnEnd)
+Process::Process(pid_t pid, bool terminateOnEnd, bool isAttached)
     : m_pid(pid)
     , m_terminateOnEnd(terminateOnEnd)
+    , m_isAttached(isAttached)
 { }
 
 } // namespace pdb
